% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_name.R
\name{clean_name}
\alias{clean_name}
\title{Nettoyage de chaine de caractères dans les noms de maille}
\usage{
clean_name(name)
}
\arguments{
\item{name}{character}
}
\value{
chaine de caractère avec une majuscule en premier, plus d espace superflu
}
\description{
Nettoyage de chaine de caractères dans les noms de maille
}
\examples{
clean_name(' il y  a des espaces    en trop! ')
}
