% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanName.R
\name{cleanName}
\alias{cleanName}
\title{Nettoyage de chaine de caractères dans les noms de maille}
\usage{
cleanName(name)
}
\arguments{
\item{name}{character}
}
\value{
chaine de caractère avec une majuscule en premier, plus d espace superflu
}
\description{
Nettoyage de chaine de caractères dans les noms de maille
}
\examples{
cleanName(' il y  a des espaces    en trop! ')
}
