% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_energie_an_maille.R
\name{get_energie_an_maille}
\alias{get_energie_an_maille}
\title{Fonction interne qui récupère l'énergie annuelle (conso ou prod) pour
une maille données (epci, nom_maille, departement, region)}
\usage{
get_energie_an_maille(
  annee,
  maille,
  nom_maille,
  prod = FALSE,
  verbose = FALSE,
  with_coord = FALSE
)
}
\arguments{
\item{annee}{entier, annee a recuperer (si non fourni, toutes)}

\item{maille}{character (epci, nom_maille, departement, region, iris) : maille d'intérêt}

\item{nom_maille}{character, nom de la maille à récupérer}

\item{prod}{booleen qui indique si on récupère la conso (FALSE, defaut) ou la prod (TRUE)}

\item{with_coord}{booleen, est ce qu'on doit récupérer les coordonnées}
}
\value{
data frame avec en ligne les données de consommation par nom_maille x segment de clientèle
}
\description{
Fonction interne qui récupère l'énergie annuelle (conso ou prod) pour
une maille données (epci, nom_maille, departement, region)
}
