% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getConsoAnCommune.R
\name{getConsoAnCommune}
\alias{getConsoAnCommune}
\title{Récupération des données de consommation électrique annuelle
d'une commune}
\usage{
getConsoAnCommune(annee_choisie, commune, with_coord = FALSE, verbose = FALSE)
}
\arguments{
\item{annee_choisie}{entier, annee a recuperer (si missing, toutes)}

\item{with_coord}{booleen, est ce qu'on doit récupérer les coordonnées}

\item{verbose}{booleen qui indique si on print l'url}
}
\value{
data frame avec en ligne les données de consommation par commune x segment de clientèle
}
\description{
Récupération des données de consommation électrique annuelle
d'une commune
}
\examples{
getConsoAnCommune(annee = 2016, commune = 'Valenciennes')
getConsoAnCommune( commune = 'paris')
}
